use typescript for vhu examples
