+++ back/src/vhu/examples/workflows/index.ts
@@ -1,0 +1,5 @@
+import vhuVersBroyeur from "./vhuVersBroyeur";
+
+export default {
+  vhuVersBroyeur
+};
